update (Client): refactor Profile page and game stats
